refactor: move classes back to their previous spot
