+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/EventListUtils.java
@@ -1,7 +1,7 @@
 package io.javaoperatorsdk.operator;
 
 import io.fabric8.kubernetes.client.Watcher;
-import io.javaoperatorsdk.operator.api.Event;
+import io.javaoperatorsdk.operator.processing.event.Event;
 import io.javaoperatorsdk.operator.processing.event.internal.CustomResourceEvent;
 import java.util.List;
