+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/api/DefaultContext.java
@@ -1,6 +1,8 @@
 package io.javaoperatorsdk.operator.api;
 
 import io.fabric8.kubernetes.client.CustomResource;
+import io.javaoperatorsdk.operator.processing.event.EventList;
+import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
 import java.util.Optional;
 
 public class DefaultContext<T extends CustomResource> implements Context<T> {
