+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/api/ResourceController.java
@@ -1,6 +1,7 @@
 package io.javaoperatorsdk.operator.api;
 
 import io.fabric8.kubernetes.client.CustomResource;
+import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
 import java.util.Locale;
 
 public interface ResourceController<R extends CustomResource> {
