+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/EventDispatcher.java
@@ -8,14 +8,14 @@ import static io.javaoperatorsdk.operator.processing.KubernetesResourceUtils.mar
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
+import io.javaoperatorsdk.operator.ControllerUtils;
 import io.javaoperatorsdk.operator.api.Context;
-import io.javaoperatorsdk.operator.api.ControllerUtils;
 import io.javaoperatorsdk.operator.api.DefaultContext;
 import io.javaoperatorsdk.operator.api.DeleteControl;
-import io.javaoperatorsdk.operator.api.EventList;
-import io.javaoperatorsdk.operator.api.EventSourceManager;
 import io.javaoperatorsdk.operator.api.ResourceController;
 import io.javaoperatorsdk.operator.api.UpdateControl;
+import io.javaoperatorsdk.operator.processing.event.EventList;
+import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
 import java.util.ArrayList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
