+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/AbstractEventSource.java
@@ -1,8 +1,5 @@
 package io.javaoperatorsdk.operator.processing.event;
 
-import io.javaoperatorsdk.operator.api.EventHandler;
-import io.javaoperatorsdk.operator.api.EventSource;
-
 public abstract class AbstractEventSource implements EventSource {
 
   protected EventHandler eventHandler;
