+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/Event.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.processing.event;
 
 public interface Event {
