+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/EventHandler.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.processing.event;
 
 public interface EventHandler {
