+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/EventList.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.processing.event;
 
 import java.util.List;
 import java.util.Optional;
