+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/EventSourceManager.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.processing.event;
 
 import java.util.Map;
 import java.util.Optional;
